Fix rating type

+++ src/db/db_models.py
@@ -72,7 +72,7 @@ class PlayerGame(DbBase):
     type: Mapped[str] = mapped_column(String(255), nullable=False)
     item_title: Mapped[str] = mapped_column(String(255), nullable=False)
     item_review: Mapped[str] = mapped_column(Text, nullable=False)
-    item_rating: Mapped[float] = mapped_column(Integer, nullable=False)
+    item_rating: Mapped[float] = mapped_column(Float, nullable=False)
     item_length: Mapped[str] = mapped_column(String(255), nullable=False)
     vod_links: Mapped[str] = mapped_column(String(255), nullable=True)
     sector_id: Mapped[int] = mapped_column(Integer, nullable=False, index=True)
